Move YouTube-related services into a separate package
