+++ src/main/java/com/ypm/controller/PlayListController.java
@@ -5,9 +5,9 @@ import com.google.api.services.youtube.model.PlaylistItem;
 import com.google.api.services.youtube.model.PlaylistSnippet;
 import com.ypm.dto.PlaylistDto;
 import com.ypm.dto.request.MergePlayListsRequest;
-import com.ypm.service.PlayListService;
+import com.ypm.service.youtube.PlayListService;
 import com.ypm.service.TokenService;
-import com.ypm.service.VideoService;
+import com.ypm.service.youtube.VideoService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.oauth2.client.OAuth2AuthorizedClient;
