+++ src/test/java/com/ypm/controller/PlayListControllerTest.java
@@ -7,8 +7,8 @@ import com.ypm.dto.PlaylistDto;
 import com.ypm.dto.request.MergePlayListsRequest;
 import com.ypm.dto.response.ExceptionResponse;
 import com.ypm.exception.PlayListNotFoundException;
-import com.ypm.service.PlayListServiceImp;
-import com.ypm.service.VideoServiceImp;
+import com.ypm.service.youtube.PlayListServiceImp;
+import com.ypm.service.youtube.VideoServiceImp;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
